single, add, running page update

+++ running_member.php
@@ -15,17 +15,12 @@
         <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.1.3/css/bootstrap.css" rel="stylesheet" id="bootstrap-css">
         <link href="https://cdn.datatables.net/1.10.19/css/dataTables.bootstrap4.min.css" rel="stylesheet" id="bootstrap-css">
 
-
+      
+        <script src="https://ajax.aspnetcdn.com/ajax/jQuery/jquery-3.4.1.min.js"></script>    
         <link href="https://fonts.maateen.me/solaiman-lipi/font.css" rel="stylesheet">
-
-        <script src="https://ajax.aspnetcdn.com/ajax/jQuery/jquery-3.4.1.min.js"></script>
-
-        <link href="css/simple-sidebar.css" rel="stylesheet">
-
-    
-
         <link href="https://use.fontawesome.com/releases/v5.0.6/css/all.css" rel="stylesheet">
         <link href="css/style.css" rel="stylesheet">
+      
 
 
 
@@ -61,69 +56,52 @@
 
         <h2 class="text-white header-top-bg text-center pt-1 pb-1">সদস্য তালিকা</h2>
 
-<div class="d-flex" id="wrapper">
-
-<!-- Sidebar -->
-<div class="bg-dark border-right" id="sidebar-wrapper">
-  <div class="sidebar-heading text-white">তরুন সংঘ একতা সমিতি লিঃ</div>
-  <div class="list-group list-group-flush">
-    <a href="#" class="list-group-item border-bottom border-top text-white list-group-item-action bg-dark"><i class="fas fa-home"></i>  মূলপাতা</a>
-    <a href="running_member.php" class="list-group-item border-bottom border-top text-white list-group-item-action bg-dark">সদস্য</a>
-    <a href="#" class="list-group-item border-bottom border-top text-white list-group-item-action bg-dark">Overview</a>
-    <a href="#" class="list-group-item border-bottom border-top text-white list-group-item-action bg-dark">Events</a>
-    <a href="#" class="list-group-item border-bottom border-top text-white list-group-item-action bg-dark">Profile</a>
-    <a href="#" class="list-group-item border-top text-white list-group-item-action bg-dark">Status</a>
-   
-  </div>
-</div>
-<!-- /#sidebar-wrapper -->
-
-
-
-<!-- Page Content -->
-<div id="page-content-wrapper">
 
 
 
 
 
 
+
+<!-- navbar -->
+<div class="container">
   <nav class="navbar navbar-expand-lg navbar-dark bg-dark border-bottom">
-    <button class="btn btn-primary btn-sm" id="menu-toggle"><<||>></button>
-    <a href="running_member.php" class="btn btn-sm btn-secondary ml-2">বর্তমান সদস্য</a>
-    <a href="running_member.php" class="btn btn-sm btn-secondary ml-2">বর্তমান সদস্য</a>
-    <a href="running_member.php" class="btn btn-sm btn-secondary ml-2">বর্তমান সদস্য</a>
-    <a href="running_member.php" class="btn btn-sm btn-secondary ml-2">বর্তমান সদস্য</a>
-    <a href="add_member.php" class="btn btn-sm btn-info ml-2"><i class="fas fa-user-plus"></i></a>
+      <button class="btn btn-primary btn-sm" id="menu-toggle"><<||>></button>
+      <a href="running_member.php" class="btn btn-sm btn-secondary ml-2">বর্তমান সদস্য</a>
+      <a href="running_member.php" class="btn btn-sm btn-secondary ml-2">বর্তমান সদস্য</a>
+      <a href="running_member.php" class="btn btn-sm btn-secondary ml-2">বর্তমান সদস্য</a>
+      <a href="running_member.php" class="btn btn-sm btn-secondary ml-2">বর্তমান সদস্য</a>
+      <a href="add_member.php" class="btn btn-sm btn-info ml-2"><i class="fas fa-user-plus"></i></a>
 
-    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-      <span class="navbar-toggler-icon"></span>
-    </button>
-    <div class="collapse navbar-collapse" id="navbarSupportedContent">
-      <ul class="navbar-nav ml-auto mt-2 mt-lg-0">
-        <li class="nav-item active">
-          <a class="nav-link" href="#">Home <span class="sr-only">(current)</span></a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="#">Link</a>
-        </li>
-        <li class="nav-item dropdown">
-          <a class="nav-link text-white dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-          সদস্য
-          </a>
-          <div class="dropdown-menu dropdown-menu-right" aria-labelledby="navbarDropdown">
-            <a class="dropdown-item" href="running_member.php">বর্তমান সদস্য</a>
-            <a class="dropdown-item" href="#">পরিশোধকৃত সদস্য</a>
-            <div class="dropdown-divider"></div>
-            <a class="dropdown-item" href="#">সদস্য</a>
-          </div>
-        </li>
-      </ul>
-    </div>
-  </nav>
+      <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+      </button>
+      <div class="collapse navbar-collapse" id="navbarSupportedContent">
+        <ul class="navbar-nav ml-auto mt-2 mt-lg-0">
+          <li class="nav-item active">
+            <a class="nav-link" href="#">Home <span class="sr-only">(current)</span></a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="#">Link</a>
+          </li>
+          <li class="nav-item dropdown">
+            <a class="nav-link text-white dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+            সদস্য
+            </a>
+            <div class="dropdown-menu dropdown-menu-right" aria-labelledby="navbarDropdown">
+              <a class="dropdown-item" href="running_member.php">বর্তমান সদস্য</a>
+              <a class="dropdown-item" href="#">পরিশোধকৃত সদস্য</a>
+              <div class="dropdown-divider"></div>
+              <a class="dropdown-item" href="#">সদস্য</a>
+            </div>
+          </li>
+        </ul>
+      </div>
+    </nav>
+</div>
 
 
-
+<!-- topbar -->
   <div class="container bg-Secondary">
     <div class="row">
         <div class="col-md-12">
@@ -132,7 +110,7 @@
                 <h5 >বর্তমান সদস্য :
                               
                 <?php 
-                $sql="SELECT * FROM member_data ORDER BY id";
+                $sql="SELECT * FROM all_member_form_data ORDER BY id";
                 if ($result=mysqli_query($conn,$sql))
                   {
                   // Return the number of rows in result set
@@ -155,7 +133,7 @@
                 </div>
             </div>
             
-            <table class="table table-striped table-dark table-bordered"  id="datatable">
+            <table class="table table-striped table-dark table-bordered" id="datatable">
                 <thead class="text-center">
                   <tr>
                     <th>তারিখ</th>
@@ -168,13 +146,12 @@
                 </tr>
                 </thead>
 
-                
 
             
                 <?php
                   
                   
-                    $sql = "SELECT id, loan_date, m_name, f_name, loan_amount, total_amount FROM member_data ORDER BY id ASC";
+                    $sql = "SELECT id, loan_date, m_name, f_name, loan_amount, total_amount FROM all_member_form_data ORDER BY id ASC";
                     $res = $conn->query($sql);
                     
                     if ($res->num_rows > 0) {
@@ -186,7 +163,7 @@
                                 <td> " . $row["loan_amount"] . "</td>
                                 <td> " . $row["total_amount"] . "</td>          
                                 <td class='text-right'><a class='btn btn-info btn-sm' id='alert' href='single_view.php?id=" .  $row["id"] . "'>দেখুন</a>
-                                <a class='btn btn-warning btn-sm' id='alert' href='delete.php?id=" .  $row["id"] . "'>ডিলিট</a><a class='btn btn-danger btn-sm ml-1' id='alert' href='delete.php?id=" .  $row["id"] . "'>ডিলিট</a></td></tr>"; 
+                                <a class='btn btn-warning btn-sm btn-delete' value='1' name='actiondelete' Onclick='return ConfirmDelete();' id='alert'  href='delete.php?id=" .  $row["id"] . "'>ডিলিট</a><a class='btn btn-danger btn-sm ml-1' id='alert' href='delete.php?id=" .  $row["id"] . "'>ডিলিট</a></td></tr>"; 
 
                       }
                     } else {
@@ -206,9 +183,13 @@
   </div>
 
 
+  
 </div>
 
- 
+
+  
+
+
 
 <!-- page-wrapper -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
@@ -222,6 +203,7 @@
     <script src="js/main.js"></script>
     <script src="js/angular.js"></script>
 
+
   
  
     
@@ -233,5 +215,46 @@
 
 
 
+
+<script>
+$(document).ready(function() {
+  $('#datatable').DataTable();
+} );
+</script>
+
+<script>
+$(document).ready(function () {
+document.querySelector('warning').onclick = function () {
+  swal({
+  title: 'Are you sure?',
+  text: 'Some text.',
+  type: 'warning',
+  showCancelButton: true,
+  confirmButtonColor: '#DD6B55',
+  confirmButtonText: 'Yes!',
+  cancelButtonText: 'No.'
+}).then(() => {
+  if (result.value) {
+    // handle Confirm button click
+  } else {
+    // result.dismiss can be 'cancel', 'overlay', 'esc' or 'timer'
+  }
+});
+    };
+
+});
+</script>
+
+
+<script>
+    function ConfirmDelete()
+    {
+      var x = confirm( "আপনি কি এই তথ্য ডিলিট করতে চান?");
+      if (x)
+          return true;
+      else
+        return false;
+    }
+</script>
